# backend/docs/swagger/complete-swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Vertex Markets API",
    "version": "1.0.0",
    "description": "Production-ready backend API for a modern Forex/Trading company website. Provides authentication, user management, CMS content management, markets, trading platforms, blog, contact management, newsletter subscriptions, careers/jobs, partner management, file uploads, admin dashboard, and global search capabilities.",
    "contact": {
      "name": "Vertex Markets Development Team"
    },
    "license": {
      "name": "MIT"
    }
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Local development server"
    },
    {
      "url": "https://api.vertexmarkets.com",
      "description": "Production server"
    }
  ],
  "security": [
    {
      "bearerAuth": []
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "JWT access token obtained from POST /api/v1/login"
      },
      "cookieAuth": {
        "type": "apiKey",
        "in": "cookie",
        "name": "accessToken",
        "description": "JWT access token stored in HTTP-only cookie"
      }
    },
    "schemas": {
...
      "PlatformCreateRequest": {
        "type": "object",
        "required": ["name", "slug", "description"],
        "properties": {
          "name": { "type": "string", "example": "MetaTrader 4" },
          "slug": { "type": "string", "example": "metatrader-4" },
          "description": { "type": "string", "example": "Industry-standard trading platform" },
          "logo": { "type": "string", "example": "mt4-logo.png" },
          "features": { "type": "array", "items": { "type": "string" } },
          "supportedMarkets": { "type": "array", "items": { "type": "string" } },
          "downloadLinks": { "type": "object" },
          "order": { "type": "integer", "example": 1 },
          "active": { "type": "boolean", "example": true }
        }
      },
...
    }
  },
  "paths": {
...
    "/api/v1/platforms": {
      "get": {
        "tags": ["Platforms"],
        "summary": "List all trading platforms",
        "description": "Retrieve a paginated list of all trading platforms. Supports filtering and search.",
        "security": [],
        "parameters": [
          { "in": "query", "name": "page", "schema": { "type": "integer", "minimum": 1 } },
          { "in": "query", "name": "limit", "schema": { "type": "integer", "minimum": 1, "maximum": 100 } },
          { "in": "query", "name": "sortBy", "schema": { "type": "string" } },
          { "in": "query", "name": "sortOrder", "schema": { "type": "string", "enum": ["asc", "desc"] } },
          { "in": "query", "name": "active", "schema": { "type": "boolean" } },
          { "in": "query", "name": "q", "schema": { "type": "string" } }
        ],
        "responses": {
          "200": { "description": "List of platforms" }
        }
      },
      "post": {
        "tags": ["Platforms"],
        "summary": "Create a trading platform (admin)",
        "description": "Create a new trading platform entry. Requires Admin role with cms:write permission.",
        "security": [{ "bearerAuth": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/PlatformCreateRequest" }
            }
          }
        },
        "responses": {
          "201": { "description": "Platform created" },
          "400": { "description": "Validation failed" },
          "401": { "description": "Unauthorized" },
          "403": { "description": "Forbidden" }
        }
      }
    },
    "/api/v1/platforms/{id}": {
      "get": {
        "tags": ["Platforms"],
        "summary": "Get platform by ID",
        "description": "Retrieve a specific trading platform by its ID.",
        "security": [],
        "parameters": [
          { "in": "path", "name": "id", "required": true, "schema": { "type": "string" } }
        ],
        "responses": {
          "200": { "description": "Platform details" },
          "404": { "description": "Platform not found" }
        }
      },
      "put": {
        "tags": ["Platforms"],
        "summary": "Update platform (admin)",
        "description": "Fully replace a trading platform entry.",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          { "in": "path", "name": "id", "required": true, "schema": { "type": "string" } }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/PlatformCreateRequest" }
            }
          }
        },
        "responses": {
          "200": { "description": "Platform updated" },
          "400": { "description": "Validation failed" },
          "401": { "description": "Unauthorized" },
          "403": { "description": "Forbidden" }
        }
      },
      "patch": {
        "tags": ["Platforms"],
        "summary": "Patch platform (admin)",
        "description": "Partially update a trading platform entry.",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          { "in": "path", "name": "id", "required": true, "schema": { "type": "string" } }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": { "type": "string" },
                  "description": { "type": "string" },
                  "logo": { "type": "string" },
                  "features": { "type": "array", "items": { "type": "string" } },
                  "supportedMarkets": { "type": "array", "items": { "type": "string" } },
                  "downloadLinks": { "type": "object" },
                  "order": { "type": "integer" },
                  "active": { "type": "boolean" }
                }
              }
            }
          }
        },
        "responses": {
          "200": { "description": "Platform patched" },
          "400": { "description": "Validation failed" },
          "401": { "description": "Unauthorized" },
          "403": { "description": "Forbidden" }
        }
      },
      "delete": {
        "tags": ["Platforms"],
        "summary": "Delete platform (admin)",
        "description": "Delete a trading platform entry.",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          { "in": "path", "name": "id", "required": true, "schema": { "type": "string" } }
        ],
        "responses": {
          "200": { "description": "Platform deleted" },
          "401": { "description": "Unauthorized" },
          "403": { "description": "Forbidden" }
        }
      }
    },
    "/api/v1/platforms/slug/{slug}": {
      "get": {
        "tags": ["Platforms"],
        "summary": "Get platform by slug",
        "description": "Retrieve a specific trading platform by its URL-friendly slug.",
        "security": [],
        "parameters": [
          { "in": "path", "name": "slug", "required": true, "schema": { "type": "string" } }
        ],
        "responses": {
          "200": { "description": "Platform details" },
          "404": { "description": "Platform not found" }
        }
      }
    },
...
  },
...
}